requirements.txt
asgiref==3.11.1
Django==6.0.2
sqlparse==0.5.5
tzdata==2025.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
whitenoise==6.8.2
